Update deploy view and allow upload via text input
