+++ src/app/frontend/common/components/uploadfile/uploadfile.html
@@ -21,7 +21,7 @@ limitations under the License.
          class="ng-hide">
   <div flex="auto">
     <md-input-container class="md-block kd-upload-file-container">
-      <label>[[Choose kubeconfig file|A 'Choose kubeconfig file' placeholder for an empty file input field on login page.]]</label>
+      <label>{{label}}</label>
       <input id="textInput"
              ng-model="fileName"
              ng-readonly="true"
