--- src/app/frontend/deploy/component.js
@@ -1,56 +1,0 @@
-// Copyright 2017 The Kubernetes Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import {deployAppStateName, deployFileStateName} from './state';
-
-/**
- * Controller for the deploy view.
- *
- * @final
- */
-class DeployController {
-  /**
-   * @param {!ui.router.$state} $state
-   * @ngInject
-   */
-  constructor($state) {
-    /** @export {string} */
-    this.selection = $state.current.name;
-
-    /** @export {string} */
-    this.appOption = deployAppStateName;
-
-    /** @export {string} */
-    this.fileOption = deployFileStateName;
-
-    /** @private {!ui.router.$state} */
-    this.state_ = $state;
-  }
-
-  /** @export */
-  changeSelection() {
-    this.state_.go(this.selection);
-  }
-}
-
-/**
- * Returns component definition for deploy component.
- *
- * @return {!angular.Component}
- */
-export const deployComponent = {
-  controller: DeployController,
-  controllerAs: 'ctrl',
-  templateUrl: 'deploy/deploy.html',
-};
