+++ src/app/frontend/deploy/createnamespace_controller.js
@@ -22,10 +22,9 @@ export default class NamespaceDialogController {
    * @param {!md.$dialog} $mdDialog
    * @param {!angular.$log} $log
    * @param {!angular.$resource} $resource
-   * TODO (cheld) Set correct type after fixing issue #159
-   * @param {!Object} errorDialog
+   * @param {!../common/errorhandling/dialog.ErrorDialog} errorDialog
    * @param {!Array<string>} namespaces
-   * @param {!./../common/csrftoken/service.CsrfTokenService} kdCsrfTokenService
+   * @param {!../common/csrftoken/service.CsrfTokenService} kdCsrfTokenService
    * @param {string} kdCsrfTokenHeader
    * @ngInject
    */
@@ -40,10 +39,7 @@ export default class NamespaceDialogController {
     /** @private {!angular.$resource} */
     this.resource_ = $resource;
 
-    /**
-     * TODO (cheld) Set correct type after fixing issue #159
-     * @private {!Object}
-     */
+    /** @private {!../common/errorhandling/dialog.ErrorDialog} */
     this.errorDialog_ = errorDialog;
 
     /**
