+++ src/app/frontend/deploy/deploy.html
@@ -15,26 +15,19 @@ limitations under the License.
 -->
 
 <kd-content-card>
-  <kd-title>[[Deploy a Containerized App|Title text which appears on top of the deploy page.]]</kd-title>
-  <kd-content class="kd-content-padded">
-    <kd-help-section>
-      <md-radio-group ng-model="ctrl.selection"
-                      ng-change="ctrl.changeSelection()">
-        <md-radio-button value="{{::ctrl.appOption}}"
-                         class="md-primary">
-          [[Specify app details below|Text for a selection option, which the user must click to manually enter the app details on the deploy page.]]
-        </md-radio-button>
-        <md-radio-button value="{{::ctrl.fileOption}}"
-                         class="md-primary">
-          [[Upload a YAML or JSON file|Text for a selection option, which the user must click to upload a YAML/JSON file to deploy from on the deploy page.]]
-        </md-radio-button>
-      </md-radio-group>
-      <kd-user-help>
-        [[To learn more, <a href="http://kubernetes.io/docs/user-guide/ui/"
-           target="_blank"
-           tabindex="-1"> take the Dashboard Tour <i class="material-icons">open_in_new</i></a>|User help with a link redirecting to the "Dashboard tour" on the deploy page.]]
-      </kd-user-help>
-    </kd-help-section>
-    <div ui-view></div>
+  <kd-content>
+    <md-tabs md-dynamic-height
+             md-border-bottom
+             class="kd-deploy-tabs">
+      <md-tab label="[[Upload from input|Upload from input tab on deploy page.]]">
+        <kd-deploy-from-input></kd-deploy-from-input>
+      </md-tab>
+      <md-tab label="[[Upload from file|Upload from file tab on deploy page.]]">
+        <kd-deploy-from-file></kd-deploy-from-file>
+      </md-tab>
+      <md-tab label="[[Specify deployment details|Specify deployment details tab on deploy page.]]">
+        <kd-deploy-from-settings></kd-deploy-from-settings>
+      </md-tab>
+    </md-tabs>
   </kd-content>
 </kd-content-card>
