+++ src/app/frontend/deploy/deploy.scss
@@ -12,23 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {UploadController} from './upload_controller';
+@import '../variables';
 
-/**
- * Provides upload file mechanism and updates the model with file parameter for the selected file.
- *
- * @return {!angular.Directive}
- * @ngInject
- */
-export default function uploadDirective() {
-  return {
-    scope: {},
-    bindToController: {
-      'file': '=',
-      'form': '=',
-    },
-    controller: UploadController,
-    controllerAs: 'ctrl',
-    templateUrl: 'deploy/upload.html',
-  };
+.kd-deploy-tabs {
+  md-tab-content {
+    padding: 2.5 * $baseline-grid;
+  }
 }
