--- src/app/frontend/deploy/deployfromfile.html
@@ -1,32 +1,0 @@
-<!--
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<form name="ctrl.form"
-      ng-submit="ctrl.deploy()"
-      novalidate>
-  <kd-upload file="ctrl.file"
-             form="ctrl.form"></kd-upload>
-  <div></div>
-  <md-button class="md-raised md-primary kd-deploy-submit-button"
-             type="submit"
-             ng-disabled="ctrl.isDeployDisabled()">
-    [[Upload|The text is put on the button at the end of the YAML upload page.]]
-  </md-button>
-  <md-button class="md-raised kd-deploy-cancel-button"
-             ng-click="ctrl.cancel()">
-    [[Cancel|The text is put on the 'Cancel' button at the end of the YAML upload page.]]
-  </md-button>
-</form>
