+++ src/app/frontend/deploy/deployfromfile/deployfromfile.scss
@@ -12,14 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/**
- * Returns a reusable directive which adds user help.
- * @return {!angular.Directive} the directive definition
- */
-export default function() {
-  return {
-    scope: {},
-    templateUrl: 'deploy/helpsection/userhelp.html',
-    transclude: true,
-  };
+@import '../../variables';
+
+.kd-deploy-from-file-info {
+  margin-bottom: $baseline-grid * 3;
+  padding: 0 (0.5 * $baseline-grid);
+
+  .material-icons {
+    font-size: $body-font-size-base;
+  }
+}
+
+.kd-deploy-from-file-button {
+  margin-left: 0;
 }
