+++ src/app/frontend/deploy/deployfromsettings/helpsection/helpsection.html
@@ -17,4 +17,5 @@ limitations under the License.
 <ng-transclude class="kd-help-section"
                layout="row"
                layout-xs="column"
-               layout-sm="column"> </ng-transclude>
+               layout-sm="column">
+</ng-transclude>
