+++ src/app/frontend/deploy/module.js
@@ -18,18 +18,21 @@ import errorHandlingModule from '../common/errorhandling/module';
 import historyModule from '../common/history/module';
 import validatorsModule from '../common/validators/module';
 
-import {deployComponent} from './component';
-import {deployFromFileComponent} from './deployfromfile_component';
-import {deployFromSettingsComponent} from './deployfromsettings_component';
+import {deployFromFileComponent} from './deployfromfile/component';
+import {deployFromInputComponent} from './deployfrominput/component';
+import {deployFromSettingsComponent} from './deployfromsettings/component';
+import helpSectionModule from './deployfromsettings/helpsection/module';
 import {deployLabelComponent} from './deploylabel_component';
 import {environmentVariablesComponent} from './environmentvariables_component';
-import fileReaderDirective from './filereader_directive';
-import helpSectionModule from './helpsection/helpsection_module';
-import initConfig from './initconfig';
 import {portMappingsComponent} from './portmappings_component';
+import {DeployService} from './service';
 import stateConfig from './stateconfig';
 import uniqueNameDirective from './uniquename_directive';
-import uploadDirective from './upload_directive';
+import {LabelKeyNameLengthValidator} from './validators/labelkeynamelengthvalidator';
+import {LabelKeyNamePatternValidator} from './validators/labelkeynamepatternvalidator';
+import {LabelKeyPrefixLengthValidator} from './validators/labelkeyprefixlengthvalidator';
+import {LabelKeyPrefixPatternValidator} from './validators/labelkeyprefixpatternvalidator';
+import {LabelValuePatternValidator} from './validators/labelvaluepatternvalidator';
 import validImageReferenceDirective from './validimagereference_directive';
 import validProtocolDirective from './validprotocol_directive';
 
@@ -45,6 +48,7 @@ export default angular
           'ngMaterial',
           'ngResource',
           'ui.router',
+          'ui.ace',
           helpSectionModule.name,
           errorHandlingModule.name,
           validatorsModule.name,
@@ -58,10 +62,23 @@ export default angular
     .directive('kdUniqueName', uniqueNameDirective)
     .directive('kdValidImagereference', validImageReferenceDirective)
     .directive('kdValidProtocol', validProtocolDirective)
-    .directive('kdFileReader', fileReaderDirective)
-    .directive('kdUpload', uploadDirective)
     .component('kdEnvironmentVariables', environmentVariablesComponent)
     .component('kdDeployLabel', deployLabelComponent)
-    .component('kdDeploy', deployComponent)
     .component('kdDeployFromFile', deployFromFileComponent)
-    .component('kdDeployFromSettings', deployFromSettingsComponent);
+    .component('kdDeployFromInput', deployFromInputComponent)
+    .component('kdDeployFromSettings', deployFromSettingsComponent)
+    .service('kdDeployService', DeployService);
+
+
+/**
+ * @param {!../common/validators/factory.ValidatorFactory} kdValidatorFactory
+ * @ngInject
+ */
+function initConfig(kdValidatorFactory) {
+  kdValidatorFactory.registerValidator('labelKeyNameLength', new LabelKeyNameLengthValidator());
+  kdValidatorFactory.registerValidator('labelKeyPrefixLength', new LabelKeyPrefixLengthValidator());
+  kdValidatorFactory.registerValidator('labelKeyNamePattern', new LabelKeyNamePatternValidator());
+  kdValidatorFactory.registerValidator(
+      'labelKeyPrefixPattern', new LabelKeyPrefixPatternValidator());
+  kdValidatorFactory.registerValidator('labelValuePattern', new LabelValuePatternValidator());
+}
