+++ src/app/frontend/deploy/stateconfig.js
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {stateName as chromeStateName} from '../chrome/state';
+import {stateName as parentState} from '../chrome/state';
 import {breadcrumbsConfig} from '../common/components/breadcrumbs/service';
 
-import {baseStateName, deployAppStateName, deployFileStateName} from './state';
+import {stateName, stateUrl} from './state';
 
 /**
  * Configures states for the deploy view.
@@ -24,34 +24,22 @@ import {baseStateName, deployAppStateName, deployFileStateName} from './state';
  * @ngInject
  */
 export default function stateConfig($stateProvider) {
-  $stateProvider.state(baseStateName, {
-    parent: chromeStateName,
-    component: 'kdDeploy',
-    abstract: true,
-    url: '/deploy',
-  });
-  $stateProvider.state(deployAppStateName, {
-    parent: baseStateName,
-    component: 'kdDeployFromSettings',
-    url: '/app',
+  $stateProvider.state(stateName, {
+    parent: parentState,
+    url: stateUrl,
     resolve: {
       'namespaceList': resolveNamespaces,
       'protocolsResource': getProtocolsResource,
       'protocolList': getDefaultProtocols,
     },
+    views: {
+      '': {
+        templateUrl: 'deploy/deploy.html',
+      },
+    },
     data: {
       [breadcrumbsConfig]: {
         'label': i18n.MSG_BREADCRUMBS_DEPLOY_APP_LABEL,
-      },
-    },
-  });
-  $stateProvider.state(deployFileStateName, {
-    parent: baseStateName,
-    component: 'kdDeployFromFile',
-    url: '/file',
-    data: {
-      [breadcrumbsConfig]: {
-        'label': i18n.MSG_BREADCRUMBS_DEPLOY_FILE_LABEL,
       },
     },
   });
@@ -93,7 +81,4 @@ const i18n = {
       app form view.
    */
   MSG_BREADCRUMBS_DEPLOY_APP_LABEL: goog.getMsg('Create an app'),
-
-  /** @type {string} @desc Breadcrum label for the YAML upload form */
-  MSG_BREADCRUMBS_DEPLOY_FILE_LABEL: goog.getMsg('Upload'),
 };
