+++ src/app/frontend/login/kubeconfig.html
@@ -15,5 +15,7 @@ limitations under the License.
 -->
 
 <div ng-if="$ctrl.selected">
-  <kd-upload-file on-load="$ctrl.onFileLoad(file)"></kd-upload-file>
+  <kd-upload-file on-load="$ctrl.onFileLoad(file)"
+                  label="[[Choose kubeconfig file|A 'Choose kubeconfig file' placeholder for an empty file input field on login page.]]">
+  </kd-upload-file>
 </div>
